refactor: split cli status helpers
